pyproject.toml
[project]
name = "crud-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = "."
